TONS of changes due to enum refactor
Updates names and locations of multiple files. Included new Nuget updates for the extension packages

+++ SharpWrapperTests/SharpWrap2534Tests/JsonConvertTests.cs
@@ -40,7 +40,7 @@ namespace SharpWrap2534Tests
             {
                 DataSize = 10,
                 ProtocolId = ProtocolId.ISO15765,
-                TxFlags = (uint)TxFlags.ISO15765_FRAME_PAD,
+                TxFlags = TxFlags.ISO15765_FRAME_PAD,
                 Data = new byte[] { 0x00, 0x00, 0x07, 0xE8, 0x41, 0x00, 0xBF, 0xFF, 0xB9, 0x93 },
             };
